Answer_Correctness

+++ DynamicProgramming/Dynamic Programming and Greedy/arrangeBuildings.java
@@ -5,20 +5,21 @@ public class arrangeBuildings {
 
   public static void main(String[] args) throws Exception {
     // write your code here
+    // We use long instead of int because for long test case like 38-45 it will fail. So, thats the change
     Scanner scn = new Scanner(System.in);
-    int n = scn.nextInt();
+    long n = scn.nextInt();
     
-    int ob = 1;
-    int os = 1;
+    long ob = 1L;
+    long os = 1L;
     
     for(int i = 2; i <= n; i++){
-        int nb = os;
-        int ns = os + ob;
+        long nb = os;
+        long ns = os + ob;
         
         os = ns;
         ob = nb;
     }
-    int total = (os + ob);
+    long total = (os + ob);
     total = total * total;
     System.out.println(total);
   }
